changeset: setup workflow, adapt package.json
The github action now takes care of all release matters:

When there are no changeset files, it'll consider the current version released.
When there are changeset files, it'll open a pull request for the version bump.
If we feel it's time for a release (with the changes indicated in the PR), we
merge that PR, and the action again takes care of publishing to github releases
and npmjs.com.

New changeset files can be created using `npx changeset`. It'll ask you for the
desired version bump (major/minor/patch), and a description of the change.
This can be done in the same PR/commit that does a change, or later,
separately.

Signed-off-by: Stephan Renatus <stephan@styra.com>